Create a project in the examples folder and move all the needed files in there
Export the FieldInfo and the Query from the index.d.ts
Update the example's Readme
