+++ examples/connection/typescript/src/employee.ts
