+++ examples/connection/typescript/tsconfig.json
